+++ app/serializers/bank_cell_serializer.rb
@@ -4,6 +4,7 @@ class BankCellSerializer < ActiveModel::Serializer
   def game_item
     return nil if object.game_item.nil?
     {
+      id: object.game_item_id,
       name: object.game_item.name,
       level: object.game_item.level,
       icon_name: object.game_item.icon_name,