+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Parser for bank data
 - render bank cells
 - BankRequest model
+- creating bank requests
 
 ## [1.0.1] - 2019-12-11
 ### Added